requirements.txt
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.53.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.0.1
Jinja2==3.1.4
kiwisolver==1.4.5
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pip==24.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
seaborn==0.13.2
setuptools==68.2.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzdata==2024.1
visitor==0.1.3
Werkzeug==3.0.3
wheel==0.41.2
WTForms==3.1.2
